chore: simplify and reduce bundle size
