+++ apps/site/types/userOS.ts
@@ -1,1 +1,1 @@
-export type UserOS = 'MAC' | 'WIN' | 'LINUX' | 'AIX' | 'OTHER' | 'LOADING';
+export type UserOS = 'MAC' | 'WIN' | 'LINUX' | 'AIX' | 'OTHER';
