+++ apps/site/util/downloadUtils.ts
@@ -115,7 +115,6 @@ export const bitnessItems = {
     },
   ],
   OTHER: [],
-  LOADING: [],
 };
 
 type formatDropdownItemsType = {
